+++ 12-2025-2026/_feladat/alkalmazasfejlesztes/feladat_2026_05_26_star_trek/StarTrek_VB/StarTrek_VB.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.14.37216.2 d17.14
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "StarTrek_VB_Console", "StarTrek_VB_Console\StarTrek_VB_Console.csproj", "{4424C9E9-95F3-4DE5-983D-7DA989C5FB55}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "StartTrek_VB_Lib", "StartTrek_VB_Lib\StartTrek_VB_Lib.csproj", "{0C70F65C-ECE2-4E23-A702-4313F0DFD017}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{4424C9E9-95F3-4DE5-983D-7DA989C5FB55}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4424C9E9-95F3-4DE5-983D-7DA989C5FB55}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4424C9E9-95F3-4DE5-983D-7DA989C5FB55}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4424C9E9-95F3-4DE5-983D-7DA989C5FB55}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0C70F65C-ECE2-4E23-A702-4313F0DFD017}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0C70F65C-ECE2-4E23-A702-4313F0DFD017}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0C70F65C-ECE2-4E23-A702-4313F0DFD017}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0C70F65C-ECE2-4E23-A702-4313F0DFD017}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {45D5B50C-23E7-473A-9A0B-F09CAF301149}
+	EndGlobalSection
+EndGlobal
