+++ 12-2025-2026/_feladat/alkalmazasfejlesztes/feladat_2026_05_26_star_trek/StarTrek_VB/StartTrek_VB_Lib/Input/hajo_osztalyok.csv
@@ -1,0 +1,137 @@
+﻿osztaly_id;osztaly_nev;szerep_id
+1;Constitution;1
+2;Defiant;2
+3;Drexia;3
+4;Brel;4
+5;Sarcophagus;5
+6;Danube;6
+7;Intrepid;7
+8;Nebula;8
+9;Excelsior;1
+10;Galaxy;1
+11;Nova;7
+12;Olympic;7
+13;Sovereign;1
+14;Ambassador;1
+15;Yorktown;1
+16;Odyssey;8
+17;Guardian;1
+18;NX;9
+19;Miranda;1
+20;Centaur;9
+21;ShiKahr;9
+22;Jupiter;8
+23;Zilant;10
+24;Balaur;8
+25;Khales;11
+26;Bortasqu;11
+27;Tulwar;8
+28;Chimera;15
+29;Haapax;10
+30;Dderidex;8
+31;Scoprion;12
+32;Scimitar;8
+33;Valdor;13
+34;Mogai;13
+35;Vesta;14
+36;Type 8;6
+37;Cube;17
+38;Sphere;17
+39;Diamond;18
+40;Tactical Cube;17
+41;Cardenas;1
+42;Malachowski;1
+43;Hoover;1
+44;Walker;1
+45;Nimitz;1
+46;Suurok;1
+47;Apolllo;7
+48;Type F;6
+49;Hermes;16
+50;Barkano;3
+51;Galor;1
+52;Ravinok;3
+53;Keldon;1
+54;Peregrine;15
+55;BokNor;3
+56;Selek;3
+57;Hideki;6
+58;Kvort;4
+59;D12;4
+60;Shvhal;1
+61;California;19
+62;Type 6A;6
+63;Galileo;6
+64;Type 7;6
+65;Type 6;6
+66;Engle;1
+67;Renaissance;16
+68;Type 15;6
+69;Constellation;1
+70;Crossfield;16
+71;Magee;16
+72;Dauntless;16
+73;Venatic;20
+74;Kumari;11
+75;Khyzon;2
+76;Charal;2
+77;Antares;3
+78;Daedalus;1
+79;Saladin;26
+80;Freedom;9
+81;Oberth;7
+82;Kelvin;14
+83;Reliant;1
+84;Shepard;11
+85;Sydney;24
+86;New Orleans;9
+87;DKora;10
+88;D7;11
+89;Raptor;16
+90;Type 10;6
+91;Type 9;6
+92;Aerowing;6
+93;Janeway;7
+94;Zhang Sui;6
+95;Tvaro;4
+96;OTV type K42;6
+97;Type 18;6
+98;Type 18H;6
+99;JemHadar;11
+100;Perikan;22
+101;Akira;1
+102;Curry;23
+103;D5;11
+104;Dama (D4);11
+105;Koloth;11
+106;Luna;7
+107;Norway;23
+108;Inquery;10
+109;Sutherland;7
+110;Syracuse;26
+111;Courage;7
+112;DKyr;1
+113;DKyr;7
+114;ShRan;1
+115;Maymora;10
+116;Spinner;1
+117;Vorcha;1
+118;Ranger;16
+119;QeHpu;27
+120;Thris;7
+121;Sevaijen;27
+122;Yravas;12
+123;Taholsin;26
+124;Tanathooef;9
+125;Thofsin;1
+126;Thozyn;2
+127;Plen Zha;3
+128;Feklhr;28
+129;Qang;29
+130;Deresus;29
+131;Chel Grett;11
+132;Rezreth;26
+133;Plesh Brek;9
+134;Brinok;1
+135;Groumall;3
+136;NeghVar;8
