+++ 12-2025-2026/_feladat/alkalmazasfejlesztes/feladat_2026_05_26_star_trek/StarTrek_VB/StartTrek_VB_Lib/Tables/HajoSzerepek.cs
@@ -1,0 +1,15 @@
+﻿namespace StartTrek_VB_Lib.Tables;
+
+internal class HajoSzerepek
+{
+    public int SzerepId { get; init; }
+    public string SzerepNev { get; init; }
+
+    internal HajoSzerepek(string dataLines)
+    {
+        string[] split = dataLines.Split(';');
+
+        SzerepId = int.Parse(split[0]);
+        SzerepNev = split[1];
+    }
+}
