--- 12-2025-2026/_feladat/alkalmazasfejlesztes/feladat_2026_05_26_startrek/StarTrek_VB/StartTrek_VB_Lib/Tables/Fajok.cs
@@ -1,15 +1,0 @@
-﻿namespace StartTrek_VB_Lib.Tables;
-
-internal class Fajok
-{
-    public int FajId { get; init; }
-    public string FajNev { get; init; }
-
-    internal Fajok(string dataLines)
-    {
-        string[] split = dataLines.Split(';');
-
-        FajId = int.Parse(split[0]);
-        FajNev = split[1];
-    }
-}
