--- 12-2025-2026/_feladat/alkalmazasfejlesztes/feladat_2026_05_26_startrek/StarTrek_VB/StartTrek_VB_Lib/Tables/HajoOsztalyok.cs
@@ -1,17 +1,0 @@
-﻿namespace StartTrek_VB_Lib.Tables;
-
-internal class HajoOsztalyok
-{
-    public int OsztalyId { get; init; }
-    public string OsztalyNev { get; init; }
-    public int SzerepId { get; init; }
-
-    internal HajoOsztalyok(string dataLines)
-    {
-        string[] split = dataLines.Split(';');
-
-        OsztalyId = int.Parse(split[0]);
-        OsztalyNev = split[1];
-        SzerepId = int.Parse(split[2]);
-    }
-}
