--- 12-2025-2026/_feladat/alkalmazasfejlesztes/feladat_2026_05_26_startrek/input/fajok.csv
@@ -1,19 +1,0 @@
-﻿faj_id;faj_nev
-15;Andoriai
-18;Bajori
-5;Borg
-13;Breen
-12;Cardassiai
-9;Egyesült Földi Csillagflotta
-16;Ferengi
-1;Föderáció
-10;Gorn
-14;Harogin
-4;Jem'Hadar
-2;Klingon
-11;Reman
-3;Romulán
-6;Talaxiai
-8;Terrai
-17;Tholian
-7;Vulkáni
