--- 12-2025-2026/_feladat/alkalmazasfejlesztes/feladat_2026_05_26_startrek/input/hajo_szerepek.csv
@@ -1,29 +1,0 @@
-﻿szerep_id;szerep_nev
-17;ismeretlen
-1;cirkáló
-11;csatacirkáló
-10;csatahajó
-8;dreadnaught
-22;elfogó
-16;felderítő hajó
-14;felfedező hajó
-9;fregatt
-18;futárhajó
-21;holohajó
-2;kísérőhajó
-23;közepes cirkáló
-7;kutatóhajó
-13;nehéz harcimadár
-15;nehéz romboló
-4;ragadozómadár
-26;romboló
-5;szarkofág
-24;személyszállító
-19;támogató hajó
-3;teherhajó
-6;űrkomp
-12;vadászgép
-20;vadászhajó
-27;könnyű cirkáló
-28;nehéz cirkáló
-29;nehéz felderítő
